pyproject.toml
[tool.poetry]
name = "redisblog"
version = "0.1.0"
description = ""
authors = ["Artemowkin <artyom.groshev2017@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
redis = "^3.5.3"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
